requirements.txt
cs50
Flask
Flask-Session
requests
datetime
werkzeug.security